third_party/WebKit/LayoutTests/bluetooth/idl-BluetoothUUIDs.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_throws(null, function() { new BluetoothUUIDs(); },
                'the constructor should not be callable with "new"');
  assert_throws(null, function() { BluetoothUUIDs(); },
                'the constructor should not be callable');
}, 'BluetoothUUIDs IDL test');
</script>
